requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
pydantic>=2.0.0

# Statistical Modeling & Bayesian Inference
pymc>=5.0.0
arviz>=0.15.0
tensorflow-probability>=0.21.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Change Point Detection
ruptures>=1.1.0

# Causal Inference
causalml>=0.15.0
dowhy>=0.11.0
networkx>=3.0.0

# Time Series Analysis
pmdarima>=2.0.0
prophet>=1.1.0

# Data Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Web APIs & Data Sources
tweepy>=4.14.0
yfinance>=0.2.0
fredapi>=0.5.0
pytrends>=4.9.0
textblob>=0.17.0

# Async & Streaming
asyncio
aiohttp>=3.8.0
websockets>=11.0.0

# Configuration & Logging
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Deployment
fastapi>=0.103.0
uvicorn>=0.23.0
gunicorn>=21.0.0

# Utilities
joblib>=1.3.0
tqdm>=4.66.0
click>=8.1.0